Implement create + validation flow for model version support in Detector
