+++ aws-frauddetector-detector/src/aws_frauddetector_detector/helpers/api_helpers.py
@@ -565,7 +565,6 @@ def call_get_external_models(frauddetector_client, model_endpoint: str = None):
     return frauddetector_client.get_external_models(**args)
 
 
-@retry_not_found_exceptions
 @api_call_with_debug_logs
 def call_get_model_version(frauddetector_client, model_id: str, model_type: str, model_version_number: str):
     args = {"modelId": model_id, "modelType": model_type, "modelVersionNumber": model_version_number}
